scripts/integration_test/requirements.txt
pytest
requests
pydantic
python-dotenv